+++ core/modules/modCamt053ReaderAndLink.class.php
@@ -55,7 +55,7 @@ class modCamt053ReaderAndLink extends DolibarrModules
 
 		$this->editor_name = 'Slordef';
 		$this->editor_url = '';
-		$this->version = '2.0.1';
+		$this->version = '2.0.2';
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
 		$this->picto = 'fa-building-columns';